+++ application-workloads/cloudera/cloudera-tableau/scripts/datagen/ddl/bin_flat/partsupp.sql
@@ -1,8 +1,0 @@
-create database if not exists ${DB};
-use ${DB};
-
-drop table if exists partsupp;
-
-create table partsupp
-stored as ${FILE}
-as select * from ${SOURCE}.partsupp;