+++ application-workloads/ibm-cloud-pak/ibm-cloud-pak-for-data/scripts/deployCloudPakDB2aaService.sh
@@ -1,0 +1,154 @@
+#!/bin/sh
+
+export SUDOUSER=$1
+export OPENSHIFTPASSWORD=$2
+export CPDNAMESPACE=$3
+export STORAGEOPTION=$4
+export ASSEMBLY=$5
+export CLUSTERNAME=$6
+export DOMAINNAME=$7
+export OPENSHIFTUSER=$8
+export APIKEY=$9
+export CHANNEL=${10}
+export VERSION=${11}
+
+export INSTALLERHOME=/home/$SUDOUSER/.ibm
+export OPERATORNAMESPACE=ibm-common-services
+export INSTALLERHOME=/home/$SUDOUSER/.ibm
+export OCPTEMPLATES=/home/$SUDOUSER/.openshift/templates
+export CPDTEMPLATES=/home/$SUDOUSER/.cpd/templates
+
+# Set parameters
+if [[ $STORAGEOPTION == "portworx" ]]; then
+    STORAGECLASS_VALUE="portworx-shared-gp3"
+    STORAGEVENDOR_VALUE="portworx"
+elif [[ $STORAGEOPTION == "ocs" ]]; then
+    STORAGECLASS_VALUE="ocs-storagecluster-cephfs"
+    STORAGEVENDOR_VALUE="ocs"
+elif [[ $STORAGEOPTION == "nfs" ]]; then
+    STORAGECLASS_VALUE="nfs"
+    STORAGEVENDOR_VALUE=""
+fi
+
+#Login
+var=1
+while [ $var -ne 0 ]; do
+echo "Attempting to login $OPENSHIFTUSER to https://api.${CLUSTERNAME}.${DOMAINNAME}:6443 "
+oc login "https://api.${CLUSTERNAME}.${DOMAINNAME}:6443" -u $OPENSHIFTUSER -p $OPENSHIFTPASSWORD --insecure-skip-tls-verify=true
+var=$?
+echo "exit code: $var"
+done
+
+
+# db2aaservice operator and CR creation 
+
+#runuser -l $SUDOUSER -c "cat > $CPDTEMPLATES/ibm-db2aaservice-catalogsource.yaml <<EOF
+#apiVersion: operators.coreos.com/v1alpha1
+#kind: CatalogSource
+#metadata:
+#  name: ibm-db2aaservice-cp4d-operator-catalog
+#  namespace: openshift-marketplace
+#spec:
+#  displayName: IBM Db2aaservice CP4D Catalog
+#  image: icr.io/cpopen/ibm-db2aaservice-cp4d-operator-catalog@sha256:a0d9b6c314193795ec1918e4227ede916743381285b719b3d8cfb05c35fec071
+#  imagePullPolicy: Always
+#  publisher: IBM
+#  sourceType: grpc
+#  updateStrategy:
+#    registryPoll:
+#      interval: 45m
+#EOF"
+
+runuser -l $SUDOUSER -c "cat > $CPDTEMPLATES/ibm-db2aaservice-sub.yaml <<EOF
+apiVersion: operators.coreos.com/v1alpha1
+kind: Subscription
+metadata:
+  name: ibm-db2aaservice-cp4d-operator
+  namespace: $OPERATORNAMESPACE
+spec:
+  channel: $CHANNEL
+  name: ibm-db2aaservice-cp4d-operator
+  installPlanApproval: Automatic
+  source: ibm-operator-catalog
+  sourceNamespace: openshift-marketplace
+EOF"
+
+runuser -l $SUDOUSER -c "cat > $CPDTEMPLATES/ibm-db2aaservice-cr.yaml <<EOF
+apiVersion: databases.cpd.ibm.com/v1
+kind: Db2aaserviceService
+metadata:
+  name: db2aaservice-cr
+  namespace: $CPDNAMESPACE
+spec:
+  storageClass: $STORAGECLASS_VALUE
+  version: \"$VERSION\"
+  license:
+    accept: true
+    license: \"Enterprise\"
+EOF"
+
+# Create Subscription. 
+
+#runuser -l $SUDOUSER -c "oc create -f $CPDTEMPLATES/ibm-db2aaservice-catalogsource.yaml"
+#runuser -l $SUDOUSER -c "echo 'Sleeping 2m for catalogsource to be created'"
+#runuser -l $SUDOUSER -c "sleep 2m"
+
+runuser -l $SUDOUSER -c "oc create -f $CPDTEMPLATES/ibm-db2aaservice-sub.yaml"
+runuser -l $SUDOUSER -c "echo 'Sleeping 2m for sub to be created'"
+runuser -l $SUDOUSER -c "sleep 2m"
+
+runuser -l $SUDOUSER -c "echo 'Sleeping 2m for operator to install'"
+runuser -l $SUDOUSER -c "sleep 2m"
+
+
+
+# Check ibm-db2aaservice-cp4d-operator-controller-manager pod status
+
+podname="ibm-db2aaservice-cp4d-operator-controller-manager"
+name_space=$OPERATORNAMESPACE
+status="unknown"
+while [ "$status" != "Running" ]
+do
+  pod_name=$(oc get pods -n $name_space | grep $podname | awk '{print $1}' )
+  ready_status=$(oc get pods -n $name_space $pod_name  --no-headers | awk '{print $2}')
+  pod_status=$(oc get pods -n $name_space $pod_name --no-headers | awk '{print $3}')
+  echo $pod_name State - $ready_status, podstatus - $pod_status
+  if [ "$ready_status" == "1/1" ] && [ "$pod_status" == "Running" ]
+  then 
+  status="Running"
+  else
+  status="starting"
+  sleep 10 
+  fi
+  echo "$pod_name is $status"
+done
+
+## Creating ibm-db2aaservice cr
+
+runuser -l $SUDOUSER -c "oc project $CPDNAMESPACE; oc create -f $CPDTEMPLATES/ibm-db2aaservice-cr.yaml"
+
+# Check CR Status
+
+SERVICE="Db2aaserviceService"
+CRNAME="db2aaservice-cr"
+SERVICE_STATUS="db2aaserviceStatus"
+
+STATUS=$(oc get $SERVICE $CRNAME -n $CPDNAMESPACE -o json | jq .status.$SERVICE_STATUS | xargs) 
+
+while  [[ ! $STATUS =~ ^(Completed|Complete)$ ]]; do
+    echo "$CRNAME is Installing!!!!"
+    sleep 60 
+    STATUS=$(oc get $SERVICE $CRNAME -n $CPDNAMESPACE -o json | jq .status.$SERVICE_STATUS | xargs) 
+    if [ "$STATUS" == "Failed" ]
+    then
+        echo "**********************************"
+        echo "$CRNAME Installation Failed!!!!"
+        echo "**********************************"
+        exit
+    fi
+done 
+echo "*************************************"
+echo "$CRNAME Installation Finished!!!!"
+echo "*************************************"
+
+echo "$(date) - ############### Script Complete #############"