+++ application-workloads/minecraft/minecraft-on-ubuntu/install_minecraft.sh
@@ -9,6 +9,7 @@
 # $7 = spawn-monsters
 # $8 = generate-structures
 # $9 = level-seed
+# $10 = location of the Minecraft server launcher
 
 # basic service and API settings
 minecraft_server_path=/srv/minecraft_server
@@ -16,8 +17,6 @@ minecraft_user=minecraft
 minecraft_group=minecraft
 UUID_URL=https://api.mojang.com/users/profiles/minecraft/$1
 
-# screen scrape the server jar location from the Minecraft server download page
-SERVER_JAR_URL="curl -L https://minecraft.net/en-us/download/server/ | grep -Eo \"(http|https)://[a-zA-Z0-9./?=_-]*\" | sort | uniq | grep launcher"
 server_jar=server.jar
 
 # add and update repos
@@ -33,10 +32,7 @@ mkdir $minecraft_server_path
 cd $minecraft_server_path
 
 # download the server jar
-while ! echo y | wget `eval $SERVER_JAR_URL`; do
-    sleep 10
-    wget `eval $SERVER_JAR_URL`
-done
+wget -O $10
 
 # set permissions on install folder
 chown -R $minecraft_user $minecraft_server_path