+++ application-workloads/opendx/openedx-tutor-lilac-ubuntu/script/deploy_tutor.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+echo "deploy tutor"
+#configure tutor
+tutor local quickstart
+
+read -p "what's your admin user email: " EMAIL
+tutor local createuser --staff --superuser admin $EMAIL
+echo "create admin user, can log in with email"
+
+