+++ application-workloads/sccm/sccm-technicalpreview/DSC/PSConfiguration.ps1
@@ -63,12 +63,18 @@ configuration Configuration
             DependsOn = "[InstallFeatureForSCCM]InstallFeature"
         }
 
+        DownloadAndInstallODBC DownloadAndInstallODBC
+        {
+            Ensure = "Present"
+            DependsOn = "[InstallADK]ADKInstall"
+        }
+
         DownloadSCCM DownLoadSCCM
         {
             CM = $CM
             ExtPath = $LogPath
             Ensure = "Present"
-            DependsOn = "[InstallADK]ADKInstall"
+            DependsOn = "[DownloadAndInstallODBC]DownloadAndInstallODBC"
         }
 
         SetDNS DnsServerAddress